Isolate some Tuya Device code in a new component.
